All distances between each pair of trees

All distances between each pair of trees

CompareAllTrees(trees, exact = FALSE, slow = TRUE, verbose = FALSE)

Arguments

trees
List of bifurcating trees of class phylo.

exact
Logical specifying whether to calculate exact rearrangement distances.

slow
Logical specifying whether to report distance measures that are slow to calculate (quartet distance and maximum agreement subtree).

verbose
Logical specifying whether to report which calculation is presently being performed.

See also

Examples

  trees <- lapply(rep(8, 5), ape::rtree, br = NULL)
  CompareAllTrees(trees)

#> $rf
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0   10   10   10   10
#> [2,]   10    0    6   10    8
#> [3,]   10    6    0   10   10
#> [4,]   10   10   10    0    8
#> [5,]   10    8   10    8    0
#> 
#> $rfi
#>              [,1]         [,2]         [,3]         [,4]         [,5]
#> [1,] 9.237056e-14 4.093417e+01 3.983463e+01 4.141959e+01 4.141959e+01
#> [2,] 4.093417e+01 9.237056e-14 2.599690e+01 4.141959e+01 3.450073e+01
#> [3,] 3.983463e+01 2.599690e+01 9.237056e-14 4.032006e+01 4.032006e+01
#> [4,] 4.141959e+01 4.141959e+01 4.032006e+01 8.526513e-14 3.498616e+01
#> [5,] 4.141959e+01 3.450073e+01 4.032006e+01 3.498616e+01 8.526513e-14
#> 
#> $ja2
#>           [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.0000000 0.7777778 0.8415556 0.8515556 0.8166667
#> [2,] 0.7777778 0.0000000 0.5360000 0.8957778 0.7111111
#> [3,] 0.8415556 0.5360000 0.0000000 0.8235556 0.7666667
#> [4,] 0.8515556 0.8957778 0.8235556 0.0000000 0.6152778
#> [5,] 0.8166667 0.7111111 0.7666667 0.6152778 0.0000000
#> 
#> $ja4
#>           [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.0000000 0.9160494 0.9723217 0.9640217 0.9430556
#> [2,] 0.9160494 0.0000000 0.5897600 0.9799109 0.7604938
#> [3,] 0.9723217 0.5897600 0.0000000 0.9649417 0.9305556
#> [4,] 0.9640217 0.9799109 0.9649417 0.0000000 0.7217496
#> [5,] 0.9430556 0.7604938 0.9305556 0.7217496 0.0000000
#> 
#> $jna2
#>           [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.0000000 0.7277778 0.8415556 0.8293333 0.7944444
#> [2,] 0.7277778 0.0000000 0.5137778 0.8513333 0.6443889
#> [3,] 0.8415556 0.5137778 0.0000000 0.8235556 0.7666667
#> [4,] 0.8293333 0.8513333 0.8235556 0.0000000 0.6027778
#> [5,] 0.7944444 0.6443889 0.7666667 0.6027778 0.0000000
#> 
#> $jna4
#>           [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.0000000 0.9035494 0.9723217 0.9615526 0.9405864
#> [2,] 0.9035494 0.0000000 0.5872909 0.9749726 0.7521234
#> [3,] 0.9723217 0.5872909 0.0000000 0.9615526 0.9271664
#> [4,] 0.9615526 0.9749726 0.9615526 0.0000000 0.7209684
#> [5,] 0.9405864 0.7521234 0.9271664 0.7209684 0.0000000
#> 
#> $dpi
#>           [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.0000000 0.7885409 0.8483395 0.8681221 0.8266045
#> [2,] 0.7885409 0.0000000 0.5871423 0.8996085 0.7424455
#> [3,] 0.8483395 0.5871423 0.0000000 0.8260867 0.7654522
#> [4,] 0.8681221 0.8996085 0.8260867 0.0000000 0.6368498
#> [5,] 0.8266045 0.7424455 0.7654522 0.6368498 0.0000000
#> 
#> $msid
#>           [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.0000000 0.3981981 0.4667386 0.4871444 0.4817831
#> [2,] 0.3981981 0.0000000 0.3726260 0.5817546 0.5087505
#> [3,] 0.4667386 0.3726260 0.0000000 0.4490800 0.4070167
#> [4,] 0.4871444 0.5817546 0.4490800 0.0000000 0.3631502
#> [5,] 0.4817831 0.5087505 0.4070167 0.3631502 0.0000000
#> 
#> $cid
#>           [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.0000000 0.6919395 0.7752027 0.8174003 0.7648516
#> [2,] 0.6919395 0.0000000 0.5269054 0.8361560 0.6755312
#> [3,] 0.7752027 0.5269054 0.0000000 0.7516378 0.6908312
#> [4,] 0.8174003 0.8361560 0.7516378 0.0000000 0.5876233
#> [5,] 0.7648516 0.6755312 0.6908312 0.5876233 0.0000000
#> 
#> $qd
#>           [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.0000000 0.6000000 0.7571429 0.7714286 0.7285714
#> [2,] 0.6000000 0.0000000 0.5285714 0.8000000 0.6285714
#> [3,] 0.7571429 0.5285714 0.0000000 0.7142857 0.6428571
#> [4,] 0.7714286 0.8000000 0.7142857 0.0000000 0.5142857
#> [5,] 0.7285714 0.6285714 0.6428571 0.5142857 0.0000000
#> 
#> $nts
#>           [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.0000000 0.5000000 0.6066667 0.6000000 0.5666667
#> [2,] 0.5000000 0.0000000 0.3733333 0.6200000 0.4700000
#> [3,] 0.6066667 0.3733333 0.0000000 0.5866667 0.5333333
#> [4,] 0.6000000 0.6200000 0.5866667 0.0000000 0.4200000
#> [5,] 0.5666667 0.4700000 0.5333333 0.4200000 0.0000000
#> 
#> $msd
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    8   12   11   11
#> [2,]    8    0    8   13    9
#> [3,]   12    8    0   11    9
#> [4,]   11   13   11    0    8
#> [5,]   11    9    9    8    0
#> 
#> $mast
#>   1 2 3 4 5
#> 1 8 6 5 4 5
#> 2 6 8 5 5 5
#> 3 5 5 8 5 5
#> 4 4 5 5 8 5
#> 5 5 5 5 5 8
#> 
#> $masti
#>           1         2         3         4         5
#> 1 13.343602  6.714246  3.906891  1.584963  3.906891
#> 2  6.714246 13.343602  3.906891  3.906891  3.906891
#> 3  3.906891  3.906891 13.343602  3.906891  3.906891
#> 4  1.584963  3.906891  3.906891 13.343602  3.906891
#> 5  3.906891  3.906891  3.906891  3.906891 13.343602
#> 
#> $nni_l
#>   1 2 3 4 5
#> 1   5 5 5 5
#> 2 5   3 5 4
#> 3 5 3   5 5
#> 4 5 5 5   4
#> 5 5 4 5 4  
#> 
#> $nni_t
#>    1  2  3  4  5
#> 1    10 10 10 10
#> 2 10     5 10  7
#> 3 10  5    10 10
#> 4 10 10 10     7
#> 5 10  7 10  7   
#> 
#> $nni_u
#>    1  2  3  4  5
#> 1    21 21 21 21
#> 2 21    13 21 18
#> 3 21 13    21 21
#> 4 21 21 21    18
#> 5 21 18 21 18   
#> 
#> $spr
#>   1 2 3 4 5
#> 1 0 2 3 2 3
#> 2 2 0 2 2 2
#> 3 3 2 0 3 3
#> 4 2 2 3 0 2
#> 5 3 2 3 2 0
#> 
#> $tbr_l
#>   1 2 3 4 5
#> 1 0 2 2 2 2
#> 2 2 0 1 2 2
#> 3 2 1 0 2 2
#> 4 2 2 2 0 2
#> 5 2 2 2 2 0
#> 
#> $tbr_u
#>   1 2 3 4 5
#> 1 0 6 6 6 6
#> 2 6 0 3 5 6
#> 3 6 3 0 6 6
#> 4 6 5 6 0 6
#> 5 6 6 6 6 0
#> 
#> $path
#>          1         2        3         4        5
#> 1 0.000000  8.944272 8.944272  9.949874 9.949874
#> 2 8.944272  0.000000 6.928203 10.344080 8.888194
#> 3 8.944272  6.928203 0.000000  9.110434 8.660254
#> 4 9.949874 10.344080 9.110434  0.000000 8.485281
#> 5 9.949874  8.888194 8.660254  8.485281 0.000000
#>