Approximate diameter of the Kendall–Collijn metric

KCDiameter() calculates the value of the Kendall & Colijn's (2016) metric distance between two pectinate trees with nTip leaves ordered in the opposite direction, which I suggest (without any attempt at a proof) may be a useful proxy for the diameter (i.e. maximum value) of the K–C metric.

KCDiameter(nTip)

Arguments

nTip
Integer specifying number of leaves.

Value

KCDiameter() returns a single numeric.

Author

Martin R. Smith (martin.smith@durham.ac.uk)

Examples

KCMax(4)

#> Error in KCMax(4): could not find function "KCMax"