'TreeDistData' is an R package that contains the functions and scripts used to compare the performance of tree distance metrics in Smith (2020).

The "Articles" link in the navigation bar lists vignettes detailing the output of analyses corresponding to the stated sections of Smith (2020). Each of these pages includes a link to the R source code, which can be used to reproduce the analytical results.

Details

The "Reference" link lists convenience functions, which facilitate calls to "TreeDist", and data objects that store the results of lengthy analyses. The code used to generate these data objects is available in the data-raw directory.

References

Smith, M.R. (2020). Information theoretic Generalized Robinson-Foulds metrics for comparing phylogenetic trees. Bioinformatics, online ahead of print.

See also