Tabulate method outputs

Helper function to tabulate sortable data in package vignettes.

.TDDTable(Table, dat, ...)

Arguments

Table
DT::DataTable() or an equivalent tabulation function.

dat
Matrix or vector of data to be tabulated, with (row) names corresponding to the methods employed.

Value

.TDDTable() produces a table plotted using Table().

Author

Martin R. Smith (martin.smith@durham.ac.uk)