Mean distances between random pairs of trees

A three-dimensional array listing the normalized distances between 1 000 random pairs of trees drawn from the uniform distribution using ape::rtree(nTip, br=NULL).

randomTreeDistances

Format

An object of class array of dimension 20 x 13 x 197.

Source

Scripts used to generate data objects are housed in the data-raw directory.

Details

Normalization is against the maximum possible value obtainable on a pair of trees of the shapes given, with the exception of the SPR distance, which is normalized against the upper bound (spr) and lower bound (sprLB) of its diameter on a tree with n leaves. The path and matching split distances are not normalized.

Rows are named with an abbreviation of the tree comparison metric:

dpi: Variation of Phylogenetic Information (Smith, forthcoming)

msid: Variation of Matching Split Information (Smith, forthcoming)

cid: Variation of Clustering Information (Smith, forthcoming)

qd: Quartet divergence (Smith, 2019)

ja2, ja4: Jaccard-Robinson-Foulds distance (Böcker et al. 2013), with k = 2, 4, with arboreal matchings enforced

jna2, jna4: JRF distance, non-arboreal matchings permitted

nts: Nye et al. tree similarity (Nye et al. 2006)

msd: Matching Split Distance (Bogdanowicz & Giaro, 2012)

mast, masti: Size and information content of maximum agreement subtree

nni_l, nni_u: Lower and upper bounds on the nearest neighbour interchange distance (Li et al. 1996)

spr: Approximate subtree prune and regraft distance, calculated using phangorn::SPR.dist

tbr_l, tbr_u: Lower and upper bounds on the tree bisection and reconnection distance, calculated using TBRDist

rf: Robinson-Foulds distance (Robinson & Foulds 1981)

rfi: Robinson-Foulds distance, splits weighted by phylogenetic information content (Smith, forthcoming)

path: Path distance (Steel & Penny 1993)

Columns list the summary statistics of calculated tree distances: the minimum (min), 1%, 5%, 10%, 25%, 50% (i.e. median), 75%, 90%, 95%, 99% percentiles, maximum (max), mean (mean) and standard deviation (sd).

The third dimension lists the number of leaves in the trees compared.

References

Allen BL, Steel MA (2001). “Subtree transfer operations and their induced metrics on evolutionary trees.” Annals of Combinatorics, 5(1), 1--15. doi: 10.1007/s00026-001-8006-8 .

Böcker S, Canzar S, Klau GW (2013). “The generalized Robinson-Foulds metric.” In Darling A, Stoye J (eds.), Algorithms in Bioinformatics. WABI 2013. Lecture Notes in Computer Science, vol 8126, 156--169. Springer, Berlin, Heidelberg. doi: 10.1007/978-3-642-40453-5_13 .

Bogdanowicz D, Giaro K (2012). “Matching split distance for unrooted binary phylogenetic trees.” IEEE/ACM Transactions on Computational Biology and Bioinformatics, 9(1), 150--160. doi: 10.1109/TCBB.2011.48 .

Li M, Tromp J, Zhang L (1996). “Some notes on the nearest neighbour interchange distance.” In Goos G, Hartmanis J, Leeuwen J, Cai J, Wong CK (eds.), Computing and Combinatorics, volume 1090, 343--351. Springer, Berlin, Heidelberg. ISBN 978-3-540-61332-9 978-3-540-68461-9, doi: 10.1007/3-540-61332-3_168 .

Nye TMW, Liò P, Gilks WR (2006). “A novel algorithm and web-based tool for comparing two alternative phylogenetic trees.” Bioinformatics, 22(1), 117--119. doi: 10.1093/bioinformatics/bti720 .

Robinson DF, Foulds LR (1981). “Comparison of phylogenetic trees.” Mathematical Biosciences, 53(1-2), 131--147. ISSN 00255564, doi: 10.1016/0025-5564(81)90043-2 .

Smith MR (2019). “Bayesian and parsimony approaches reconstruct informative trees from simulated morphological datasets.” Biology Letters, 15, 20180632. doi: 10.1098/rsbl.2018.0632 .

Smith MR (2020). “Information theoretic Generalized Robinson-Foulds metrics for comparing phylogenetic trees.” In review.

Steel MA, Penny D (1993). “Distributions of tree comparison metrics—some new results.” Systematic Biology, 42(2), 126--141. ISSN 1076-836X, doi: 10.1093/sysbio/42.2.126 .