All distances between each pair of trees

All distances between each pair of trees

CompareAllTrees(trees, exact = FALSE, slow = TRUE, verbose = FALSE)

Arguments

trees
List of bifurcating trees of class phylo.

exact
Logical specifying whether to calculate exact rearrangement distances.

slow
Logical specifying whether to report distance measures that are slow to calculate (quartet distance and maximum agreement subtree).

verbose
Logical specifying whether to report which calculation is presently being performed.

See also

Author

Martin R. Smith (martin.smith@durham.ac.uk)

Examples

trees <- lapply(rep(8, 5), TreeTools::RandomTree, root = TRUE)
CompareAllTrees(trees)

#> $pid
#>           1         2         3         4
#> 2 0.8638710                              
#> 3 0.8765556 0.9393481                    
#> 4 0.8295972 0.9132010 0.5815350          
#> 5 0.8446827 0.7062995 0.7776109 0.6929858
#> 
#> $msid
#>           1         2         3         4
#> 2 0.5571818                              
#> 3 0.5945095 0.6140312                    
#> 4 0.5403198 0.4960348 0.3057217          
#> 5 0.4871444 0.3824559 0.4106123 0.3434586
#> 
#> $cid
#>           1         2         3         4
#> 2 0.8065354                              
#> 3 0.8070752 0.8766895                    
#> 4 0.7611890 0.8195411 0.5358147          
#> 5 0.7893264 0.6319386 0.7244807 0.6291396
#> 
#> $qd
#>           [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.0000000 0.7571429 0.7000000 0.7285714 0.7142857
#> [2,] 0.7571429 0.0000000 0.8142857 0.7571429 0.5714286
#> [3,] 0.7000000 0.8142857 0.0000000 0.3857143 0.6285714
#> [4,] 0.7285714 0.7571429 0.3857143 0.0000000 0.5000000
#> [5,] 0.7142857 0.5714286 0.6285714 0.5000000 0.0000000
#> 
#> $nye
#>           1         2         3         4         5
#> 1 0.0000000 0.5800000 0.5833333 0.5533333 0.5733333
#> 2 0.5800000 0.0000000 0.6366667 0.5866667 0.4533333
#> 3 0.5833333 0.6366667 0.0000000 0.4200000 0.5566667
#> 4 0.5533333 0.5866667 0.4200000 0.0000000 0.4666667
#> 5 0.5733333 0.4533333 0.5566667 0.4666667 0.0000000
#> 
#> $jnc2
#>           1         2         3         4
#> 2 0.8557778                              
#> 3 0.8737778 0.9360000                    
#> 4 0.8291111 0.9040000 0.6068889          
#> 5 0.8335556 0.6680000 0.7790556 0.6943889
#> 
#> $jnc4
#>           1         2         3         4
#> 2 0.9591109                              
#> 3 0.9553738 0.9897600                    
#> 4 0.9428738 0.9846400 0.7266605          
#> 5 0.9529047 0.7698800 0.9215405 0.8771234
#> 
#> $jco2
#>           1         2         3         4
#> 2 0.8113333                              
#> 3 0.8041667 0.8610556                    
#> 4 0.7846667 0.8195000 0.6068889          
#> 5 0.8026667 0.6457778 0.7790556 0.6943889
#> 
#> $jco4
#>           1         2         3         4
#> 2 0.9541726                              
#> 3 0.9422743 0.9766605                    
#> 4 0.9379356 0.9579387 0.7266605          
#> 5 0.9453156 0.7674109 0.9215405 0.8771234
#> 
#> $ms
#>    1  2  3  4
#> 2 12         
#> 3 10 12      
#> 4 12 14  8   
#> 5 11 11 11  9
#> 
#> $mast
#>   1 2 3 4 5
#> 1 8 5 5 5 5
#> 2 5 8 5 4 5
#> 3 5 5 8 5 5
#> 4 5 4 5 8 5
#> 5 5 5 5 5 8
#> 
#> $masti
#>           1         2         3         4         5
#> 1 13.343602  3.906891  3.906891  3.906891  3.906891
#> 2  3.906891 13.343602  3.906891  1.584963  3.906891
#> 3  3.906891  3.906891 13.343602  3.906891  3.906891
#> 4  3.906891  1.584963  3.906891 13.343602  3.906891
#> 5  3.906891  3.906891  3.906891  3.906891 13.343602
#> 
#> $nni_l
#>   1 2 3 4 5
#> 1   5 5 5 5
#> 2 5   5 5 4
#> 3 5 5   4 5
#> 4 5 5 4   5
#> 5 5 4 5 5  
#> 
#> $nni_L
#>    1  2  3  4  5
#> 1    10 10 10 10
#> 2 10    10 10  7
#> 3 10 10     6 10
#> 4 10 10  6    10
#> 5 10  7 10 10   
#> 
#> $nni_t
#>    1  2  3  4  5
#> 1    10 10 10 10
#> 2 10    10 10  7
#> 3 10 10     6 10
#> 4 10 10  6    10
#> 5 10  7 10 10   
#> 
#> $nni_U
#>    1  2  3  4  5
#> 1    10 10 10 10
#> 2 10    10 10  7
#> 3 10 10     6 10
#> 4 10 10  6    10
#> 5 10  7 10 10   
#> 
#> $nni_u
#>    1  2  3  4  5
#> 1    18 18 18 18
#> 2 18    18 18 14
#> 3 18 18    10 18
#> 4 18 18 10    18
#> 5 18 14 18 18   
#> 
#> $spr
#>   1 2 3 4 5
#> 1 0 2 2 2 3
#> 2 2 0 3 3 2
#> 3 2 3 0 3 2
#> 4 2 3 3 0 2
#> 5 3 2 2 2 0
#> 
#> $tbr_l
#>   1 2 3 4 5
#> 1 0 2 2 2 2
#> 2 2 0 2 2 2
#> 3 2 2 0 2 2
#> 4 2 2 2 0 2
#> 5 2 2 2 2 0
#> 
#> $tbr_u
#>   1 2 3 4 5
#> 1 0 6 6 6 5
#> 2 6 0 6 6 5
#> 3 6 6 0 6 6
#> 4 6 6 6 0 6
#> 5 5 5 6 6 0
#> 
#> $rf
#>    1  2  3  4
#> 2 10         
#> 3 10 10      
#> 4 10 10  8   
#> 5 10  8 10 10
#> 
#> $icrf
#>          1        2        3        4
#> 2 43.00456                           
#> 3 40.93417 43.00456                  
#> 4 43.00456 45.07494 32.91577         
#> 5 41.41959 36.57112 41.41959 43.48998
#> 
#> $path
#>           1        2         3         4         5
#> 1  0.000000 10.77033  9.165151 10.770330  9.746794
#> 2 10.770330  0.00000 11.135529 12.489996 10.344080
#> 3  9.165151 11.13553  0.000000  7.745967  9.110434
#> 4 10.770330 12.49000  7.745967  0.000000  8.888194
#> 5  9.746794 10.34408  9.110434  8.888194  0.000000
#>