All distances between each pair of trees

All distances between each pair of trees

CompareAllTrees(trees, exact = FALSE, slow = TRUE, verbose = FALSE)

Arguments

trees
List of bifurcating trees of class phylo.

exact
Logical specifying whether to calculate exact rearrangement distances.

slow
Logical specifying whether to report distance measures that are slow to calculate (quartet distance and maximum agreement subtree).

verbose
Logical specifying whether to report which calculation is presently being performed.

See also

Examples

trees <- lapply(rep(8, 5), TreeTools::RandomTree, root = TRUE)
CompareAllTrees(trees)

#> $pid
#>           [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.0000000 0.6855723 0.8701237 0.8916623 0.7791881
#> [2,] 0.6855723 0.0000000 0.7992712 0.8729826 0.6835717
#> [3,] 0.8701237 0.7992712 0.0000000 0.8324976 0.9004839
#> [4,] 0.8916623 0.8729826 0.8324976 0.0000000 0.7752843
#> [5,] 0.7791881 0.6835717 0.9004839 0.7752843 0.0000000
#> 
#> $msid
#>           [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.0000000 0.4097590 0.4960348 0.5663605 0.4330209
#> [2,] 0.4097590 0.0000000 0.5060460 0.4943478 0.3610478
#> [3,] 0.4960348 0.5060460 0.0000000 0.4690449 0.4892358
#> [4,] 0.5663605 0.4943478 0.4690449 0.0000000 0.4386708
#> [5,] 0.4330209 0.3610478 0.4892358 0.4386708 0.0000000
#> 
#> $cid
#>           [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.0000000 0.6244219 0.8133189 0.8268418 0.7275111
#> [2,] 0.6244219 0.0000000 0.7289253 0.8093488 0.6023676
#> [3,] 0.8133189 0.7289253 0.0000000 0.7345701 0.7733499
#> [4,] 0.8268418 0.8093488 0.7345701 0.0000000 0.7172942
#> [5,] 0.7275111 0.6023676 0.7733499 0.7172942 0.0000000
#> 
#> $qd
#>           [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.0000000 0.5428571 0.7285714 0.7571429 0.6285714
#> [2,] 0.5428571 0.0000000 0.7000000 0.7714286 0.6000000
#> [3,] 0.7285714 0.7000000 0.0000000 0.7000000 0.7000000
#> [4,] 0.7571429 0.7714286 0.7000000 0.0000000 0.6142857
#> [5,] 0.6285714 0.6000000 0.7000000 0.6142857 0.0000000
#> 
#> $nye
#>           [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.0000000 0.4466667 0.5866667 0.6066667 0.5333333
#> [2,] 0.4466667 0.0000000 0.5533333 0.6066667 0.4533333
#> [3,] 0.5866667 0.5533333 0.0000000 0.5133333 0.5133333
#> [4,] 0.6066667 0.6066667 0.5133333 0.0000000 0.5033333
#> [5,] 0.5333333 0.4533333 0.5133333 0.5033333 0.0000000
#> 
#> $jnc2
#>           [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.0000000 0.6875000 0.8680000 0.8860000 0.7832778
#> [2,] 0.6875000 0.0000000 0.7971111 0.8637778 0.6540000
#> [3,] 0.8680000 0.7971111 0.0000000 0.8280000 0.8888889
#> [4,] 0.8860000 0.8637778 0.8280000 0.0000000 0.7763889
#> [5,] 0.7832778 0.6540000 0.8888889 0.7763889 0.0000000
#> 
#> $jnc4
#>           [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.0000000 0.8552373 0.9698800 0.9772600 0.9166296
#> [2,] 0.8552373 0.0000000 0.9377538 0.9747909 0.7721400
#> [3,] 0.9698800 0.9377538 0.0000000 0.9490800 0.9580247
#> [4,] 0.9772600 0.9747909 0.9490800 0.0000000 0.8947434
#> [5,] 0.9166296 0.7721400 0.9580247 0.8947434 0.0000000
#> 
#> $jco2
#>           [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.0000000 0.6750000 0.8195556 0.8415000 0.7666667
#> [2,] 0.6750000 0.0000000 0.7846111 0.8415556 0.6457778
#> [3,] 0.8195556 0.7846111 0.0000000 0.7557778 0.7446667
#> [4,] 0.8415000 0.8415556 0.7557778 0.0000000 0.7221667
#> [5,] 0.7666667 0.6457778 0.7446667 0.7221667 0.0000000
#> 
#> $jco4
#>           [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.0000000 0.8544560 0.9589017 0.9713587 0.9158484
#> [2,] 0.8544560 0.0000000 0.9369726 0.9723217 0.7674109
#> [3,] 0.9589017 0.9369726 0.0000000 0.9341109 0.9171356
#> [4,] 0.9713587 0.9723217 0.9341109 0.0000000 0.8871543
#> [5,] 0.9158484 0.7674109 0.9171356 0.8871543 0.0000000
#> 
#> $ms
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    8   14   14   11
#> [2,]    8    0   12   14    9
#> [3,]   14   12    0   10    9
#> [4,]   14   14   10    0   11
#> [5,]   11    9    9   11    0
#> 
#> $mast
#>   1 2 3 4 5
#> 1 8 6 4 5 5
#> 2 6 8 5 4 5
#> 3 4 5 8 6 6
#> 4 5 4 6 8 5
#> 5 5 5 6 5 8
#> 
#> $masti
#>           1         2         3         4         5
#> 1 13.343602  6.714246  1.584963  3.906891  3.906891
#> 2  6.714246 13.343602  3.906891  1.584963  3.906891
#> 3  1.584963  3.906891 13.343602  6.714246  6.714246
#> 4  3.906891  1.584963  6.714246 13.343602  3.906891
#> 5  3.906891  3.906891  6.714246  3.906891 13.343602
#> 
#> $nni_l
#>   1 2 3 4 5
#> 1   5 5 5 5
#> 2 5   5 5 4
#> 3 5 5   5 5
#> 4 5 5 5   5
#> 5 5 4 5 5  
#> 
#> $nni_t
#>    1  2  3  4  5
#> 1    10 10 10 10
#> 2 10    10 10  7
#> 3 10 10    10 10
#> 4 10 10 10    10
#> 5 10  7 10 10   
#> 
#> $nni_u
#>    1  2  3  4  5
#> 1    21 21 21 21
#> 2 21    21 21 18
#> 3 21 21    21 21
#> 4 21 21 21    21
#> 5 21 18 21 21   
#> 
#> $spr
#>   1 2 3 4 5
#> 1 0 2 2 2 2
#> 2 2 0 1 3 2
#> 3 2 1 0 1 2
#> 4 2 3 1 0 2
#> 5 2 2 2 2 0
#> 
#> $tbr_l
#>   1 2 3 4 5
#> 1 0 2 2 2 2
#> 2 2 0 2 2 1
#> 3 2 2 0 2 2
#> 4 2 2 2 0 2
#> 5 2 1 2 2 0
#> 
#> $tbr_u
#>   1 2 3 4 5
#> 1 0 6 6 6 6
#> 2 6 0 6 6 3
#> 3 6 6 0 5 5
#> 4 6 6 5 0 6
#> 5 6 3 5 6 0
#> 
#> $rf
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0   10   10   10   10
#> [2,]   10    0   10   10    8
#> [3,]   10   10    0   10   10
#> [4,]   10   10   10    0   10
#> [5,]   10    8   10   10    0
#> 
#> $icrf
#>              [,1]         [,2]         [,3]         [,4]         [,5]
#> [1,] 9.237056e-14 4.300456e+01 4.507494e+01 4.507494e+01 4.348998e+01
#> [2,] 4.300456e+01 9.237056e-14 4.300456e+01 4.300456e+01 3.450073e+01
#> [3,] 4.507494e+01 4.300456e+01 9.237056e-14 4.507494e+01 4.348998e+01
#> [4,] 4.507494e+01 4.300456e+01 4.507494e+01 9.237056e-14 4.348998e+01
#> [5,] 4.348998e+01 3.450073e+01 4.348998e+01 4.348998e+01 8.526513e-14
#> 
#> $path
#>           1         2        3        4        5
#> 1  0.000000  9.380832 12.32883 12.49000 10.34408
#> 2  9.380832  0.000000 10.77033 11.13553  8.42615
#> 3 12.328828 10.770330  0.00000 11.83216 10.72381
#> 4 12.489996 11.135529 11.83216  0.00000 10.14889
#> 5 10.344080  8.426150 10.72381 10.14889  0.00000
#>