Method parameters

Metadata for methods examined in this package.

tdAbbrevs

tdPlotSequence

tdMdAbbrevs

tdBoxAbbrevs

tdMethods

tdCol

TDFunctions

TDPair

Format

An object of class expression of length 25.

Source

Scripts used to generate data objects are housed in the data-raw directory.

Details

tdAbbrevs lists abbreviations for each method, using expressions to allow formatting of text when plotted.

tdPlotSequence lists the 20 methods discussed in the main article, in the sequence in which they are plotted in figures.

tdMdAbbrevs uses markdown formatting.

tdBoxAbbrevs uses line breaks to fit abbreviations in an approximately square bounding box.

tdCol provides each method with a suitable plotting colour.

TDFunctions lists for each method a function that will calculate the distance between two trees or lists of trees.

TDPair lists for each method a function to calculate the distance between one tree (tr) and another tree (ref).