Mean distances between random pairs of trees

A three-dimensional array listing the normalized distances between 1000 random pairs of trees drawn from the uniform distribution using ape::rtree(nTip, br=NULL).

randomTreeDistances

Format

An object of class array of dimension 10 x 13 x 197.

Details

Normalization is against the maximum possible value obtainable on a pair of trees of the shapes given, with the exception of the SPR distance, which is normalized against the upper bound (spr) and lower bound (sprLB) of its diameter on a tree with n tips. The path and matching split distances are not normalized.

Rows are named with an abbreviation of the tree comparison metric:

vpi: Variation of Phylogenetic Information (Smith, forthcoming)

vmsi: Variation of Matching Split Information (Smith, forthcoming)

vci: Variation of Clustering Information (Smith, forthcoming)

qd: Quartet divergence (Smith, 2019)

nts: Nye et al. tree similarity (Nye et al. 2006)

msd: Matching Split Distance (Bogdanowicz & Giaro, 2012)

rf: Robinson-Foulds distance (Robinson & Foulds 1981)

path: Path distance (Steel & Penny 1993)

spr: SPR distance, normalized against upper bound of diameter (Allen & Steel, 2001)

sprLB: SPR distance, normalized against lower bound of diameter (Allen & Steel, 2001)

Columns list the summary statistics of calculated tree distances: the minimum (min), 1 maximum (max), mean (mean) and standard deviation (sd).

The third dimension lists the number of tips in the trees compared.

References

Allen BL, Steel MA (2001). “Subtree transfer operations and their induced metrics on evolutionary trees.” Annals of Combinatorics, 5(1), 1--15. ISSN 0218-0006, doi: 10.1007/s00026-001-8006-8 .

Bogdanowicz D, Giaro K (2012). “Matching split distance for unrooted binary phylogenetic trees.” IEEE/ACM Transactions on Computational Biology and Bioinformatics, 9(1), 150--160. doi: 10.1109/TCBB.2011.48 .

Nye TMW, Liò P, Gilks WR (2006). “A novel algorithm and web-based tool for comparing two alternative phylogenetic trees.” Bioinformatics, 22(1), 117--119. doi: 10.1093/bioinformatics/bti720 .

Robinson DF, Foulds LR (1981). “Comparison of phylogenetic trees.” Mathematical Biosciences, 53(1-2), 131--147. ISSN 00255564, doi: 10.1016/0025-5564(81)90043-2 .

Smith MR (2019). “Bayesian and parsimony approaches reconstruct informative trees from simulated morphological datasets.” Biology Letters, 15, 20180632. doi: 10.1098/rsbl.2018.0632 .

Smith MR (2019). “Information theoretic Generalized Robinson-Foulds metrics for comparing phylogenetic trees.” Forthcoming.

Steel MA, Penny D (1993). “Distributions of tree comparison metrics—some new results.” Systematic Biology, 42(2), 126--141. ISSN 1076-836X, doi: 10.1093/sysbio/42.2.126 .